update test script

+++ test_runner.py
@@ -10,34 +10,46 @@ boost_dir=os.environ.get("BOOST_PATH")
 if not boost_dir:
     boost_dir='D:\\boost_1_60_0'
 
-lua_versions = ["lua-5.3.3","lua-5.2.4","lua-5.1.5"]
+lua_versions = ["lua-5.3.3","lua-5.2.4","lua-5.1.5","luajit"]
+maijor_test_lua_versions = ["lua-5.3.3","luajit"]
 
-test_msvc_vers = [("msvc2015","Visual Studio 14 2015","")
-    ,("msvc2015win64","Visual Studio 14 2015 Win64","")
-    ,("msvc2013","Visual Studio 12 2013","")
-    ,("msvc2013win64","Visual Studio 12 2013 Win64","")
-    ,("msvc2012","Visual Studio 11 2012","-DADDITIONAL_INCLUDE_PATH="+boost_dir)
-    ,("msvc2010","Visual Studio 10 2010","-DADDITIONAL_INCLUDE_PATH="+boost_dir)
-    ,("msvc2008","Visual Studio 9 2008","-DADDITIONAL_INCLUDE_PATH="+boost_dir)]
+test_msvc_vers = [("msvc2015","Visual Studio 14 2015","",True)
+    ,("msvc2015win64","Visual Studio 14 2015 Win64","",True)
+    ,("msvc2013","Visual Studio 12 2013","",False)
+    ,("msvc2013win64","Visual Studio 12 2013 Win64","",False)
+    ,("msvc2012","Visual Studio 11 2012","-DADDITIONAL_INCLUDE_PATH="+boost_dir,False)
+    ,("msvc2010","Visual Studio 10 2010","-DADDITIONAL_INCLUDE_PATH="+boost_dir,False)
+    ,("msvc2008","Visual Studio 9 2008","-DADDITIONAL_INCLUDE_PATH="+boost_dir,True)]
 
 test_compilers = [
-    ('gcc-3.3','g++-3.3','')
-    ,('gcc-4.4','g++-4.4','')
-    ,('gcc-4.5','g++-4.5','')
-    ,('gcc-4.6','g++-4.6','-DCMAKE_CXX_FLAGS=-std=c++03')
-    ,('gcc-4.7','g++-4.7','-DCMAKE_CXX_FLAGS=-std=c++03')
-    ,('gcc-4.7','g++-4.7','-DCMAKE_CXX_FLAGS=-std=c++11')
-    ,('gcc-4.8','g++-4.8','-DCMAKE_CXX_FLAGS=-std=c++03')
-    ,('gcc-4.8','g++-4.8','-DCMAKE_CXX_FLAGS=-std=c++11')
-    ,('gcc-4.9','g++-4.9','-DCMAKE_CXX_FLAGS=-std=c++03')
-    ,('gcc-4.9','g++-4.9','-DCMAKE_CXX_FLAGS=-std=c++11')
-    ,('gcc-5','g++-5','-DCMAKE_CXX_FLAGS=-std=c++03')
-    ,('gcc-5','g++-5','-DCMAKE_CXX_FLAGS=-std=c++11')
-    ,('gcc-6','g++-6','-DCMAKE_CXX_FLAGS=-std=c++03')
-    ,('gcc-6','g++-6','-DCMAKE_CXX_FLAGS=-std=c++11')
-    ,('gcc','g++','')
-    ,('clang','clang++','-DCMAKE_CXX_FLAGS=-std=c++03')
-    ,('clang','clang++','-DCMAKE_CXX_FLAGS=-std=c++11')]
+    ('gcc-3.3','g++-3.3','',False)
+    ,('gcc-4.4','g++-4.4','',False)
+    ,('gcc-4.5','g++-4.5','',False)
+    ,('gcc-4.6','g++-4.6','-DCMAKE_CXX_FLAGS=-std=c++03',False)
+    ,('gcc-4.7','g++-4.7','-DCMAKE_CXX_FLAGS=-std=c++03',False)
+    ,('gcc-4.7','g++-4.7','-DCMAKE_CXX_FLAGS=-std=c++11',False)
+    ,('gcc-4.8','g++-4.8','-DCMAKE_CXX_FLAGS=-std=c++03',False)
+    ,('gcc-4.8','g++-4.8','-DCMAKE_CXX_FLAGS=-std=c++11',False)
+    ,('gcc-4.9','g++-4.9','-DCMAKE_CXX_FLAGS=-std=c++03',False)
+    ,('gcc-4.9','g++-4.9','-DCMAKE_CXX_FLAGS=-std=c++11',False)
+    ,('gcc-5','g++-5','-DCMAKE_CXX_FLAGS=-std=c++03',False)
+    ,('gcc-5','g++-5','-DCMAKE_CXX_FLAGS=-std=c++11',False)
+    ,('gcc-6','g++-6','-DCMAKE_CXX_FLAGS=-std=c++03',False)
+    ,('gcc-6','g++-6','-DCMAKE_CXX_FLAGS=-std=c++11',False)
+    ,('gcc','g++','-DCMAKE_CXX_FLAGS=-std=c++03',True)
+    ,('gcc','g++','-DCMAKE_CXX_FLAGS=-std=c++11',True)
+    ,('clang','clang++','-DCMAKE_CXX_FLAGS=-std=c++03',True)
+    ,('clang','clang++','-DCMAKE_CXX_FLAGS=-std=c++11',True)
+    ,('clang-3.8','clang++-3.8','-DCMAKE_CXX_FLAGS=-std=c++03',False)
+    ,('clang-3.9','clang++-3.9','-DCMAKE_CXX_FLAGS=-std=c++03',False)
+    ,('clang-4.0','clang++-4.0','-DCMAKE_CXX_FLAGS=-std=c++03',False)
+    ,('clang-3.8','clang++-3.8','-DCMAKE_CXX_FLAGS=-std=c++11',False)
+    ,('clang-3.9','clang++-3.9','-DCMAKE_CXX_FLAGS=-std=c++11',False)
+    ,('clang-4.0','clang++-4.0','-DCMAKE_CXX_FLAGS=-std=c++11',False)
+    ,('clang-3.8','clang++-3.8','-DCMAKE_CXX_FLAGS=-std=c++14',False)
+    ,('clang-3.9','clang++-3.9','-DCMAKE_CXX_FLAGS=-std=c++14',False)
+    ,('clang-4.0','clang++-4.0','-DCMAKE_CXX_FLAGS=-std=c++14',False)
+    ]
 
 def build_and_exec_test(compiler,lua_version,build_type,dir_opt):
     ccompiler = compiler[0]
@@ -59,8 +71,11 @@ def build_and_exec_test(compiler,lua_version,build_type,dir_opt):
 
 def build_with_target_compiler(lua_version):
     for i,compiler in enumerate(test_compilers):
+        if not compiler[3] and lua_version not in maijor_test_lua_versions:
+            continue
         build_and_exec_test(compiler,lua_version,"Debug",str(i))
-        build_and_exec_test(compiler,lua_version,"Release",str(i))
+        if compiler[3]:
+            build_and_exec_test(compiler,lua_version,"Release",str(i))
 
 def build_msvc_and_exec_test(msvcver,lua_version,build_type):
     buildpath = '_build/'+msvcver[0]+'_'+lua_version
@@ -77,13 +92,16 @@ def build_msvc_and_exec_test(msvcver,lua_version,build_type):
 
 def build_with_msvc_ver(lua_version):
     for msvcver in test_msvc_vers:
+        if not msvcver[3] and lua_version not in maijor_test_lua_versions:
+            continue
         build_msvc_and_exec_test(msvcver,lua_version,'Debug')
-        build_msvc_and_exec_test(msvcver,lua_version,'Release')
+        if msvcver[3]:
+            build_msvc_and_exec_test(msvcver,lua_version,'Release')
 
 
 
-for luaversion in lua_versions:
-    if not os.path.exists(luaversion):
+for i,luaversion in enumerate(lua_versions):
+    if not os.path.exists(luaversion) and luaversion != 'luajit':
         if not os.path.exists(luaversion+".tar.gz"):
             lua_url = "https://www.lua.org/ftp/"+luaversion+".tar.gz"
             urllib.urlretrieve(lua_url,"{0}".format(luaversion+".tar.gz"))
@@ -91,6 +109,6 @@ for luaversion in lua_versions:
         tf.extractall("./")
 
     if os.name == 'nt':
-        build_with_msvc_ver(luaversion)
+        build_with_msvc_ver(luaversion,)
     else:
         build_with_target_compiler(luaversion)
